+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.2.0
 
+- Add genotyping variables
+- The CSV datasets generated are located here:
+  - `\\rfawin.partners.org\bwh-sleepepi-nsrr-staging\20231018-mignot-apoe\nsrr-prep\_releases\0.2.0.pre`
+
 ## 0.1.0 (January 11, 2024)
 
 - Add initial variable set and metadata from ApoE team