+++ CHANGELOG.md
@@ -1,9 +1,9 @@
-## 0.2.0
+## 0.2.0 (March 1, 2024)
 
 - Add genotyping variables
 - Fix nsrr_sex representation
 - The CSV datasets generated are located here:
-  - `\\rfawin.partners.org\bwh-sleepepi-nsrr-staging\20231018-mignot-apoe\nsrr-prep\_releases\0.2.0.pre`
+  - `\\rfawin.partners.org\bwh-sleepepi-nsrr-staging\20231018-mignot-apoe\nsrr-prep\_releases\0.2.0`
 
 ## 0.1.0 (January 11, 2024)